convert notification model to typescript
